feat: coup2stl logic for coup buildings feat: add generalized script for geojson2stl
